[bug/57] Change header components
