+++ src/layout/header/space.jsx
@@ -6,9 +6,9 @@ const useStyles = makeStyles(theme => ({
     width: 0,
     [theme.breakpoints.only('xs')]: { width: 200 },
     [theme.breakpoints.only('sm')]: { width: 200 },
-    [theme.breakpoints.only('md')]: { width: 400 },
-    [theme.breakpoints.only('lg')]: { width: 1300 },
-    [theme.breakpoints.only('xl')]: { width: 1900 },
+    [theme.breakpoints.only('md')]: { width: 100 },
+    [theme.breakpoints.only('lg')]: { width: 300 },
+    [theme.breakpoints.only('xl')]: { width: 500 },
   },
 }))
